# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pulsar_data_aggregation_API",
        "description": "",
        "contact": {
            "email": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suggestions": {
            "get": {
                "tags": [
                    "Suggestions"
                ],
                "operationId": "f4b4bb1653bec66c0964387d10039f54",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "tags": [
                    "Suggestions"
                ],
                "operationId": "047f70bcfc7e15f9f9001fdb9e42f6d7",
                "requestBody": {
                    "request": "Create Suggestions",
                    "description": "Create Suggestions Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "fio": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    },
                                    "user_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "category_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "subdivision_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "file_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "description": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "fio": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        },
                                        "user_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "category_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "subdivision_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "file_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "description": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suggestions/get-by-id/{id}": {
            "get": {
                "tags": [
                    "Suggestions"
                ],
                "operationId": "f6dcfe9fb6ff83e2fbe9032740bcf94a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "1",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suggestions/{id}": {
            "delete": {
                "tags": [
                    "Suggestions"
                ],
                "operationId": "865d51e4b0b014101ce5521b2f8731d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error Delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_deleted": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Suggestions"
                ],
                "operationId": "3e94b81494c922d4a60e57a8b2a639b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Update Suggestions",
                    "description": "Update Suggestions Fields",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "fio": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    },
                                    "user_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "0"
                                    },
                                    "category_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "subdivision_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "file_id": {
                                        "description": "0",
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "description": {
                                        "description": "0",
                                        "type": "string",
                                        "example": "Текст"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "fio": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        },
                                        "user_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "category_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "subdivision_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "file_id": {
                                            "description": "0",
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "description": {
                                            "description": "0",
                                            "type": "string",
                                            "example": "Текст"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "errors",
                                    "description": "errors object",
                                    "properties": {
                                        "errors": {
                                            "title": "Validation errors",
                                            "description": "Validation errors object",
                                            "properties": {
                                                "field1": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field1 field is required."
                                                    }
                                                },
                                                "field2": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The field2 field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}